Prevent keystore password from leaking in logs (#173)
Fixes #164.

Instead of storing the password in an env var, now we write it to a file and pass it to zookeeper by file name. I used the previous mechanism to create a unique password in each pod and then write that password to a file. This solution is backwards compatible.

Alternatively, I could have made it possible to configure this password for each component. I think that would end up being more effort than it's worth, and it would have meant that the secret was the same for all pods in a group.