Use FQDN for service names; add value for kubernetesClusterDomain (#187)
* Use FQDN for service names

* Make cluster domain configurable

* Clean up missed hard coded suffixes

* Remove last hard coded references to cluster.local

Fixes: #184.

Hostname verification relies on the initial DNS named targeted. In order to support stricter requirements with DNS hostnames, we'll default to using the fully qualified service names when configuring Pulsar components.

Note: because the DNS names are part of the cluster initialization, you'll need to create a new cluster to take advantage of this feature or you'll need to run the admin command to update the cluster: `bin/pulsar-admin clusters update ...`.