Add option to monitor pulsar components via PodMonitor (#270)
* Add option to monitor pulsar components via PodMonitor

* Add missed conditional deployment

* Fix incorrect filename

* Remove old scrape definition; improve PodMonitor

* Document the PodMonitor

* Add newline at end of file to pass linter

## Motivation

The current Prometheus Scrape Config is defined by passing it to the sub prometheus chart. As a result, it requires manual intervention for users with their own Prometheus instance to configure metrics scraping. In order to accommodate these users, I propose adding a single pod monitor that targets all of our metrics endpoints. In order to make this change, I needed to update some port names to `http`. I don't believe that will be a problem for most use cases, but it could break any that are already using the port name.

## Changes

* Add `PodMonitor` to target 9 components. These are the components with `prometheus.io/scrape: "true"`, so the end result will be the same components being monitored. The relabel configs are similar to, but not the same as the scrape config currently used: https://github.com/datastax/pulsar-helm-chart/blob/a93a7cf4beb4a83290f4a19baa57b80b64d0392e/helm-chart-sources/pulsar/values.yaml#L2085-L2108
* Add `http` port definition for autorecovery, bookkeeper, zookeeper, and zookeepernp
* Replace port names `function` and `functiontls` with `http` and `https`, respectively, on the function pod. These pod names are not used anywhere in the chart, but they could be used by some users. This breaking change should be documented in the release notes.